First attempt at a pipeline
Largely copied from Kai's ekyc pipeline:

https://bitbucket.org/openid/ekyc-ida/src/master/bitbucket-pipelines.yml

The main change is just to match the name of our spec, but I also
disabled the clone of the repo in the publish step, as it's not needed.

Note that the build step does NOT run for merge requests from forks;
this appears to be a bitbucket limitation:

https://jira.atlassian.com/browse/BCLOUD-13162

Only tries FAPI 2 Baseline spec for now. It will verify the spec
builds without errors for any branches, and for master it will push
the specs to the drafts website.

I'm not exactly sure how the ekyc group gave the pipeline write access
to the website repo, I generated a key for the FAPI WG repo pipelines
using the bitbucket ui on
https://bitbucket.org/openid/fapi/admin/addon/admin/pipelines/ssh-keys
and then added the public key to a newly created 'OpenID FAPI Pipeline
Bot' user, and granted that user write access to the website repo:

https://bitbucket.org/openid/openid.bitbucket.io

The FAPI2-Baseline draft is now available at:

https://openid.bitbucket.io/fapi/fapi-2_0-baseline.html

Part of #400