Generate html for all specs in pipeline
Add a python wrapper that generates html for all the specs that work
with the current tooling, and uploads them to the openid.bitbucket.io
site, and also generates an index.html listing the specs.

The drafts are accessible from:

https://openid.bitbucket.io/fapi/

There's an exclusion list for specs that are currently known not to
build, which Unfortunately has basically every spec on it except the
fapi2 specs. Hopefully some of these can be fixed.