FAPI2 Baseline: Align the chapter etc. structure to FAPI 1
FAPI 1.0 is using the ISO and IEC defined structure, apply the same to
FAPI 2.0 Baseline.

I used FAPI1-Adv final as a reference:

https://openid.net/specs/openid-financial-api-part-2-1_0.html

I have not been able to figure out how to get the 'normative
references' section to appear in the same place as FAPI1. Either mmark
or xml2rfc seem to generate it at the end of the document and I'm not
sure how to change that.

part of #465