Remove mention of signing id_tokens in comparison table
As per earlier clauses, unsigned id tokens are not permitted (the
'none' alg is not allowed).

Note that the situation is that the id_token returned in the
backchannel must be signed by the server, but as it is TLS protected
the client is not required to verify the signature as per OpenID
Connect core.

closes #470