FAPI2MS: Make [security profile] a real reference
The link used for the security profile is updated to the one from:

https://openid.net/2022/11/14/public-review-period-for-two-proposed-fapi-2-0-second-implementers-drafts/

The version is used as it is a stable version. Ideally we would link
to the second implementer's draft but the vote won't finish till 23rd
Jan as per:

https://openid.net/2023/01/08/notice-of-vote-for-proposed-second-implementers-drafts-of-two-fapi-2-0-specifications/

The attacker model link is updated to the corresponding version for
the same reason.

The anchor name of the reference is updated to try and make the
generated html clearer (note that space is not a valid character for
an anchor name, so _ is used instead).

relates to #560